Added due_months input validation

+++ apps/tasks/schemas.py
@@ -30,7 +30,7 @@ class ReqIn(Schema):
 class ReqOut(Schema):
     pk: int
     is_active: bool
-    due_months: float | None = None
+    due_months: int | None = None
     due_months_unit: str | None = None
     tol_pos_mos: float | None = None
     tol_neg_mos: float | None = None
